requirements.txt
# Core dependencies
fastapi==0.122.0
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
google-genai==1.52.0
python-multipart==0.0.20

# Logging
python-json-logger==4.0.0

# File validation
filetype==1.2.0

# Testing dependencies
pytest==9.0.1
pytest-asyncio==1.3.0
hypothesis==6.148.3
pytest-cov==7.0.0
httpx==0.28.1
pytest-mock==3.15.1
